Add files via upload
This is a simple implementation of Flappy Bird Game , built using Python. The objective of the game is to control a bird and navigate it through a series of pipes without touching them. The player has to tap/click space bar to make the bird fly and avoid obstacles, trying to get the highest possible score.